- Added NEXT_PUBLIC_SITE_PASSWORD to env example - Update stream form and plaground/try to use new comfyui editor - Add comfyui badge to stream list view - Remove x-user-id header requirement from webhook
