+++ apps/app/.env.example
@@ -41,3 +41,4 @@ NEXT_PUBLIC_RTMP_URL="rtmp://rtmp.livepeer.monster/live/"
 NEXT_PUBLIC_LIVEPEER_STUDIO_API_KEY=""
 NEXT_PUBLIC_LIVEPEER_STUDIO_API_URL="https://livepeer.studio.server/api"
 NEXT_PUBLIC_INTERCOM_APP_ID=""
+NEXT_PUBLIC_SITE_PASSWORD=""
